Implement fractional skill ranks for cross-class skills
Cross-class skills now cost 1 skill point per 0.5 ranks, matching D&D 3.5e rules.

## Changes
- **Schema updates**: Allow decimal values for skill ranks in Level and SkillRanks schemas
- **SkillSpendingPanel**:
  - Class skills: +/- 1 rank per point (unchanged)
  - Cross-class skills: +/- 0.5 ranks per point (new)
  - Each click costs 1 skill point regardless of class type
- **SkillsPanel**: Only full ranks (Math.floor) count toward skill modifiers
- **Visual indicator**: Shows "(half ranks don't count)" when fractional ranks present
- **CSS**: Added styling for fractional rank indicator text

## D&D 3.5e Rule Implementation
Per PHB p.62: "Cross-class skills cost 2 skill points per rank"
- Implementation: 1 point = 0.5 ranks for cross-class skills
- Modifier calculation: Only complete ranks add to skill bonus
- Example: 2.5 ranks = +2 modifier (half rank doesn't count)

+++ src/types/level.ts
@@ -11,7 +11,7 @@ export const LevelSchema = z
     level: z.number().int().min(1).max(20),
     class: z.enum(CLASS_NAMES),
     feats: z.array(z.string()).optional().default([]),
-    skillRanks: z.record(z.string(), z.number().int().min(0).max(99)).optional().default({}),
+    skillRanks: z.record(z.string(), z.number().min(0).max(99)).optional().default({}), // Allow decimals for 0.5 ranks
     unspentSkillPoints: z.number().int().min(0).optional().default(0),
   })
   .strict();
